requirements.txt
discord.py>=2.3.0
anthropic>=0.40.0
openai>=1.50.0
google-generativeai>=0.8.0
pyyaml>=6.0
python-dotenv>=1.0.0
